RESEARCH_ORCHESTRATOR/requirements.txt
requests
pyyaml
networkx
matplotlib
sentence_transformers
chromadb